Move transaction actions to lib/transaction
