+++ client/lib/transaction/action-types.js
@@ -1,0 +1,6 @@
+export const TRANSACTION_DOMAIN_DETAILS_SET = 'TRANSACTION_DOMAIN_DETAILS_SET';
+export const TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET = 'TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET';
+export const TRANSACTION_PAYMENT_SET = 'TRANSACTION_PAYMENT_SET';
+export const TRANSACTION_RESET = 'TRANSACTION_RESET';
+export const TRANSACTION_STEP_SET = 'TRANSACTION_STEP_SET';
+export const TRANSACTION_STRIPE_SET = 'TRANSACTION_STRIPE_SET';
