+++ client/lib/upgrades/action-types.js
@@ -45,12 +45,6 @@ export const SITE_REDIRECT_NOTICE_CLOSE = 'SITE_REDIRECT_NOTICE_CLOSE';
 export const SITE_REDIRECT_UPDATE = 'SITE_REDIRECT_UPDATE';
 export const SITE_REDIRECT_UPDATE_COMPLETED = 'SITE_REDIRECT_UPDATE_COMPLETED';
 export const SITE_REDIRECT_UPDATE_FAILED = 'SITE_REDIRECT_UPDATE_FAILED';
-export const TRANSACTION_DOMAIN_DETAILS_SET = 'TRANSACTION_DOMAIN_DETAILS_SET';
-export const TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET = 'TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET';
-export const TRANSACTION_PAYMENT_SET = 'TRANSACTION_PAYMENT_SET';
-export const TRANSACTION_RESET = 'TRANSACTION_RESET';
-export const TRANSACTION_STEP_SET = 'TRANSACTION_STEP_SET';
-export const TRANSACTION_STRIPE_SET = 'TRANSACTION_STRIPE_SET';
 export const WAPI_DOMAIN_INFO_FETCH = 'WAPI_DOMAIN_INFO_FETCH';
 export const WAPI_DOMAIN_INFO_FETCH_COMPLETED = 'WAPI_DOMAIN_INFO_FETCH_COMPLETED';
 export const WAPI_DOMAIN_INFO_FETCH_FAILED = 'WAPI_DOMAIN_INFO_FETCH_FAILED';
