+++ client/lib/upgrades/actions/checkout.js
@@ -1,4 +1,3 @@
-/** @format */
 /**
  * External dependencies
  */
@@ -7,53 +6,13 @@ import { defer } from 'lodash';
 /**
  * Internal dependencies
  */
-import {
-	TRANSACTION_DOMAIN_DETAILS_SET,
-	TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET,
-	TRANSACTION_PAYMENT_SET,
-	TRANSACTION_RESET,
-	TRANSACTION_STEP_SET,
-	TRANSACTION_STRIPE_SET,
-} from 'lib/upgrades/action-types';
-import Dispatcher from 'dispatcher';
+import { setTransactionStep } from 'lib/transaction/actions';
 import { submit } from 'lib/store-transactions';
-
-export function setDomainDetails( domainDetails ) {
-	Dispatcher.handleViewAction( {
-		type: TRANSACTION_DOMAIN_DETAILS_SET,
-		domainDetails,
-	} );
-}
-
-export function setPayment( payment ) {
-	Dispatcher.handleViewAction( {
-		type: TRANSACTION_PAYMENT_SET,
-		payment,
-	} );
-}
-
-export function setStripeObject( stripe, stripeConfiguration ) {
-	Dispatcher.handleViewAction( {
-		type: TRANSACTION_STRIPE_SET,
-		stripe,
-		stripeConfiguration,
-	} );
-}
-
-export function setNewCreditCardDetails( options ) {
-	const { rawDetails, maskedDetails } = options;
-
-	Dispatcher.handleViewAction( {
-		type: TRANSACTION_NEW_CREDIT_CARD_DETAILS_SET,
-		rawDetails,
-		maskedDetails,
-	} );
-}
 
 export function submitTransaction( { cart, transaction, successUrl, cancelUrl } ) {
 	submit(
 		{
-			cart: cart,
+			cart,
 			payment: transaction.payment,
 			domainDetails: transaction.domainDetails,
 			successUrl,
@@ -64,18 +23,6 @@ export function submitTransaction( { cart, transaction, successUrl, cancelUrl } 
 		// Execute every step handler in its own event loop tick, so that a complete React
 		// rendering cycle happens on each step and `componentWillReceiveProps` of objects
 		// like the `TransactionStepsMixin` are called with every step.
-		step =>
-			defer( () => {
-				Dispatcher.handleViewAction( {
-					type: TRANSACTION_STEP_SET,
-					step,
-				} );
-			} )
+		step => defer( () => setTransactionStep( step ) )
 	);
-}
-
-export function resetTransaction() {
-	Dispatcher.handleViewAction( {
-		type: TRANSACTION_RESET,
-	} );
 }
