+++ client/lib/upgrades/actions/index.js
@@ -24,14 +24,7 @@ export {
 	showCartOnMobile,
 } from './cart';
 
-export {
-	resetTransaction,
-	setDomainDetails,
-	setNewCreditCardDetails,
-	setPayment,
-	setStripeObject,
-	submitTransaction,
-} from './checkout';
+export { submitTransaction } from './checkout';
 
 export {
 	acceptTransfer,
